+++ message.json
@@ -1,0 +1,1 @@
+[{"id":"0379c574-3c1d-4c76-b694-589290f709a7","title":"Joi — awesome code validation for Node.js and Express","url":"https://dev.to/itnext/joi-awesome-code-validation-for-node-js-and-express-35pk","keywords":["tutorial","javascript","validation"],"modifiedAt":"06/28/2019","publishedAt":"06/28/2019","author":"Chris Noring","readMins":9,"source":"ARTICLE"},{"id":"4a71a08a-b31e-427e-84f3-127a1273ffd2","title":"Understanding Node.js File System Module","url":"https://levelup.gitconnected.com/understanding-node-js-file-system-module-b16da1e01949","keywords":["Nodejs","Filesystem"],"modifiedAt":"06/24/2019","publishedAt":"06/24/2019","author":"Swathi Prasad","readMins":3,"source":"BLOG"},{"id":"6e3beecd-9bbf-4ec1-8960-f15c73c87d99","title":"A Short Summary On Clean Coding Best Practices","url":"https://codeburst.io/a-short-summary-on-clean-coding-best-practices-d8afbfa7677","keywords":["Software Development","Coding","Clean Code"],"modifiedAt":"02/15/2021","publishedAt":"02/15/2021","author":"Thilina Ashen Gamage","readMins":9,"source":"BLOG"},{"id":"b7af2583-451c-41cc-9377-87296e1bf4a4","title":"Node.js Streams: Everything you need to know","url":"https://medium.com/edge-coders/node-js-streams-everything-you-need-to-know-c9141306be93","keywords":["Nodejs","Filesystem","Streams"],"modifiedAt":"05/23/2017","publishedAt":"05/23/2017","author":"Samer Buna","readMins":16,"source":"ARTICLE"},{"id":"e62f8d18-412f-4d86-8a19-99b6c3381741","title":"fs.readFile vs streams to read text files in node.js","url":"https://medium.com/@dalaidunc/fs-readfile-vs-streams-to-read-text-files-in-node-js-5dd0710c80ea","keywords":["Nodejs","Filesystem","Comparison"],"modifiedAt":"08/22/2017","publishedAt":"08/22/2017","author":"Duncan Grant","readMins":3,"source":"ARTICLE"},{"id":"e693b52e-c274-43a8-b38a-9675c8b2ac04","title":"Understanding Node.js File System basic ops for beginners.","url":"https://medium.com/@akhilgoyal_/understanding-node-js-file-system-basic-ops-for-beginners-4560ee6ebc2b","keywords":["JavaScript","Nodejs","Filesystem"],"modifiedAt":"05/11/2019","publishedAt":"05/11/2019","author":"Akhil Goyal","readMins":5,"source":"ARTICLE"}]
